+++ src/components/logo/Logo.jsx
@@ -1,0 +1,11 @@
+import {Link} from "react-router-dom"
+
+export default function Logo(){
+    return(
+        <div className="logo">
+            <Link to = "/">
+                <img src ={process.env.PUBLIC_URL + "img/logo.svg"} alt="logo"/> 
+            </Link>
+        </div>
+    )  
+}
